refactor(offline): split engine and offline UI modules below static-analysis limits (tr-8h4)
Codacy ACTION_REQUIRED on PR #230 flagged two oversized files and six
methods over the 50-LOC / complexity-10 limits:

- src/offline/engine.rs (1369 NCLOC) -> engine/{mod,machine,board}.rs
  plus engine/tests/{mod,admission,integrity,lifecycle,resume_quota}.rs
- src/ui/offline.rs (566 NCLOC) -> ui/offline/{mod,panel,tests}.rs
- step_receive (95 LOC, cc 15) -> receive_plan / apply_resume_discipline /
  write_received_segment
- board (54 LOC) -> board_keys / cached_view_for / row_snapshot
- enforce_quota_then_publish (81 LOC, cc 11) -> measure_received /
  evict_until_fits / publish_snapshot
- plan (54 LOC) -> base_accessible / append_suffix / progress helpers
- classify (79 LOC) -> classify_state -> classify_in_flight
- build_row (81 LOC) -> row_grid / heading / title_label / status_line /
  progress_bar / action_button

No behavior change: cargo fmt, clippy -D warnings, and the full test
suite (1810+26+20+1) pass at this commit.
